Move SW activation modal to a separate component
